move da-mock-service to its own module
