+++ rollup-interface/src/state_machine/mocks/da.rs
@@ -1,8 +1,6 @@
 use core::fmt::Display;
 use core::str::FromStr;
 
-#[cfg(all(feature = "native", feature = "tokio"))]
-use async_trait::async_trait;
 use borsh::{BorshDeserialize, BorshSerialize};
 use bytes::Bytes;
 use serde::{Deserialize, Serialize};
@@ -11,12 +9,8 @@ use crate::da::{
     BlobReaderTrait, BlockHashTrait, BlockHeaderTrait, CountedBufReader, DaSpec, DaVerifier, Time,
 };
 use crate::maybestd::string::String;
-#[cfg(all(feature = "native", feature = "tokio"))]
-use crate::maybestd::sync::Arc;
 use crate::maybestd::vec::Vec;
 use crate::mocks::MockValidityCond;
-#[cfg(all(feature = "native", feature = "tokio"))]
-use crate::services::da::DaService;
 use crate::services::da::SlotData;
 use crate::{BasicAddress, RollupAddress};
 
@@ -302,82 +296,6 @@ impl DaSpec for MockDaSpec {
     type InclusionMultiProof = [u8; 32];
     type CompletenessProof = ();
     type ChainParams = ();
-}
-
-#[cfg(all(feature = "native", feature = "tokio"))]
-use tokio::sync::mpsc::{self, Receiver, Sender};
-#[cfg(all(feature = "native", feature = "tokio"))]
-use tokio::sync::Mutex;
-
-#[cfg(all(feature = "native", feature = "tokio"))]
-#[derive(Clone)]
-/// DaService used in tests.
-pub struct MockDaService {
-    sender: Sender<Vec<u8>>,
-    receiver: Arc<Mutex<Receiver<Vec<u8>>>>,
-    sequencer_da_address: MockAddress,
-}
-
-#[cfg(all(feature = "native", feature = "tokio"))]
-impl MockDaService {
-    /// Creates a new MockDaService.
-    pub fn new(sequencer_da_address: MockAddress) -> Self {
-        let (sender, receiver) = mpsc::channel(100);
-        Self {
-            sender,
-            receiver: Arc::new(Mutex::new(receiver)),
-            sequencer_da_address,
-        }
-    }
-}
-
-#[cfg(all(feature = "native", feature = "tokio"))]
-#[async_trait]
-impl DaService for MockDaService {
-    type Spec = MockDaSpec;
-    type Verifier = MockDaVerifier;
-    type FilteredBlock = MockBlock;
-    type Error = anyhow::Error;
-
-    async fn get_finalized_at(&self, _height: u64) -> Result<Self::FilteredBlock, Self::Error> {
-        let data = self.receiver.lock().await.recv().await;
-        let data = data.unwrap();
-        let hash = [0; 32];
-
-        let blob = MockBlob::new(data, self.sequencer_da_address, hash);
-
-        Ok(MockBlock {
-            blobs: vec![blob],
-            ..Default::default()
-        })
-    }
-
-    async fn get_block_at(&self, height: u64) -> Result<Self::FilteredBlock, Self::Error> {
-        self.get_finalized_at(height).await
-    }
-
-    fn extract_relevant_blobs(
-        &self,
-        block: &Self::FilteredBlock,
-    ) -> Vec<<Self::Spec as DaSpec>::BlobTransaction> {
-        block.blobs.clone()
-    }
-
-    async fn get_extraction_proof(
-        &self,
-        _block: &Self::FilteredBlock,
-        _blobs: &[<Self::Spec as DaSpec>::BlobTransaction],
-    ) -> (
-        <Self::Spec as DaSpec>::InclusionMultiProof,
-        <Self::Spec as DaSpec>::CompletenessProof,
-    ) {
-        ([0u8; 32], ())
-    }
-
-    async fn send_transaction(&self, blob: &[u8]) -> Result<(), Self::Error> {
-        self.sender.send(blob.to_vec()).await.unwrap();
-        Ok(())
-    }
 }
 
 /// The configuration for mock da
